extract Breadcrumb::ItemComponent
