+++ app/components/primer/breadcrumb_component.rb
@@ -5,7 +5,12 @@ module Primer
   class BreadcrumbComponent < Primer::Component
     include ViewComponent::Slotable
 
-    with_slot :item, collection: true, class_name: "BreadcrumbItem"
+    # _Note: if both `href` and `selected: true` are passed in, `href` will be ignored and the item will not be rendered as a link._
+    #
+    # @param href [String] The URL to link to.
+    # @param selected [Boolean] Whether or not the item is selected and not rendered as a link.
+    # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
+    renders_many :items, Primer::Breadcrumb::ItemComponent
 
     # @example auto|Basic
     #   <%= render(Primer::BreadcrumbComponent.new) do |component| %>
@@ -23,24 +28,6 @@ module Primer
 
     def render?
       items.any?
-    end
-
-    # _Note: if both `href` and `selected: true` are passed in, `href` will be ignored and the item will not be rendered as a link._
-    class BreadcrumbItem < Primer::Slot
-      attr_reader :href, :system_arguments
-
-      # @param href [String] The URL to link to.
-      # @param selected [Boolean] Whether or not the item is selected and not rendered as a link.
-      # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
-      def initialize(href: nil, selected: false, **system_arguments)
-        @href = href
-        @system_arguments = system_arguments
-
-        @href = nil if selected
-        @system_arguments[:tag] = :li
-        @system_arguments[:"aria-current"] = "page" if selected
-        @system_arguments[:classes] = "breadcrumb-item #{@system_arguments[:classes]}"
-      end
     end
   end
 end
